added Patient, Countries and Referral Channel Models

+++ database/migrations/2022_01_12_112322_create_contact_details_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('contact_details', function (Blueprint $table) {
             $table->id();
-            $table->integer('type');//ContactDetailTypeEnum
+            $table->unsignedTinyInteger('type');//ContactDetailTypeEnum
             $table->text('description');
             $table->text('details');
             $table->morphs('contactable');
